[refactor] split CurrencySelector from CreditorBlock
