+++ frontend/src/pages/Group/pages/Expenses/components/CreditorsBlock.jsx
@@ -2,6 +2,7 @@ import { Container, Form, Col, Row } from "react-bootstrap";
 import { GroupContext } from "../../../../../contexts/GroupContext";
 import { ExpenseContext } from "../../../../../contexts/ExpenseContext";
 import { CURRENCY_OPTIONS } from "../../../../../global/constant";
+import CurrencySelector from "./CurrencySelector";
 import { useContext } from "react";
 import {
     List,
@@ -11,34 +12,8 @@ import {
     ListItemAvatar,
     Avatar,
     TextField,
-    MenuItem,
 } from "@mui/material";
 import { amountFormatter } from "../../../../../utils/formatter";
-
-const CurrencySelector = () => {
-    const { selectedCurrency, setSelectedCurrency } =
-        useContext(ExpenseContext);
-    const handleCurrencyOptionChange = (event) => {
-        setSelectedCurrency(event.target.value);
-    };
-
-    return (
-        <TextField
-            name="currency_option"
-            select
-            label="Currency"
-            variant="standard"
-            value={selectedCurrency}
-            onChange={handleCurrencyOptionChange}
-        >
-            {CURRENCY_OPTIONS.map((option) => (
-                <MenuItem key={option.id} value={option.id}>
-                    {option.abbreviation}
-                </MenuItem>
-            ))}
-        </TextField>
-    );
-};
 
 const CreditorsBlock = () => {
     const { members } = useContext(GroupContext);
